==> 05_Validacion_encuestas.qmd <==
# 05_Validacion_encuestas.R {-}

Para ejecutar este archivo, es necesario abrir el archivo **05_Validacion_encuestas.R** ubicado en la ruta *Rcodes/2020/05_Validacion_encuestas.R*. Este script está diseñado para llevar a cabo un análisis integral de datos con varias fases que aseguran una evaluación rigurosa de la información disponible.

En la primera etapa del código, se eliminan todos los objetos del entorno de R para garantizar que el análisis se realice en un entorno limpio y sin interferencias. Luego, se cargan las librerías necesarias para llevar a cabo el análisis y se establece el tema de visualización predeterminado usando `bayesplot::theme_default()`. También se carga un archivo de script adicional que incluye funciones específicas para la creación de gráficos. Posteriormente, se leen las bases de datos `encuesta_sta` y `muestra_ampliada`, que contienen los datos de la encuesta y del censo, respectivamente.

El análisis procede realizando comparaciones entre los datos censales y los datos de la encuesta en diferentes variables categóricas, tales como edad, sexo, nivel educativo y estado. Para cada una de estas variables, se generan gráficos que permiten visualizar las diferencias. Utilizando la librería `patchwork`, se combinan estos gráficos en un solo diseño para facilitar la comparación. Además, se llevan a cabo análisis adicionales en variables como lengua indígena y discapacidad, y se exploran las interacciones entre variables como sexo y edad, nivel educativo y sexo, y estado y sexo. Finalmente, los gráficos que ilustran estas interacciones se integran en un diseño consolidado para una visualización comprensiva de los efectos.


#### Limpieza del Entorno y Carga de Bibliotecas{-}

Se realiza una limpieza del entorno en R para eliminar todos los objetos y liberar memoria, asegurando que no haya interferencias de variables o datos previos en el análisis actual. Posteriormente, se cargan las bibliotecas necesarias para la manipulación de datos y la creación de gráficos. Entre las bibliotecas cargadas se encuentran `tidyverse` para la manipulación y visualización de datos, `reshape2` para la reorganización de datos, `stringr` para el manejo de cadenas de texto, `ggalt` para gráficos avanzados, `gridExtra` para la disposición de múltiples gráficos, `scales` para la adaptación de escalas en los gráficos, `formatR` para el formateo de código, y `patchwork` para combinar múltiples gráficos.

Se establece un tema predeterminado para los gráficos mediante `bayesplot::theme_default()` para mantener una presentación visual consistente. Además, se carga un archivo de script adicional, `Plot_validacion.R`, que contiene funciones personalizadas para la validación de gráficos, permitiendo así una integración eficiente y consistente de las herramientas gráficas necesarias para el análisis.

```{r, eval=FALSE}
#### Cleaning R environment ###
rm(list = ls())

#################
#### Libraries ###
#################
library(tidyverse)
library(reshape2)
library(stringr)
library(ggalt)
library(gridExtra)
library(scales)
library(formatR)
library(patchwork)

theme_set(bayesplot::theme_default())

source(file = "../source/Plot_validacion.R", encoding = "UTF-8")
```

#### Lectura de Datos {-}

Se leen las bases de datos necesarias para el análisis. La primera base de datos, `encuesta_sta`, se carga desde el archivo `../input/2020/enigh/encuesta_sta.rds`, y la segunda, `muestra_ampliada`, se carga desde el archivo `../input/2020/muestra_ampliada/muestra_cuestionario_ampliado.rds`. Estos conjuntos de datos proporcionan la información requerida para las siguientes etapas del análisis.

```{r, eval=FALSE}
encuesta_sta <- readRDS("../input/2020/enigh/encuesta_sta.rds")
muestra_ampliada <- readRDS("../input/2020/muestra_ampliada/muestra_cuestionario_ampliado.rds")
```

#### Comparación de Variables {-}

Se comparan las distribuciones de diferentes variables entre los datos censales y los de la encuesta mediante la función `Plot_Compare`. 

Para analizar la variable **edad**, se crea el gráfico `age_plot`, que compara las distribuciones de edad en los datos censales y en los de la encuesta. 

Se genera el gráfico `sex_plot` para comparar las distribuciones de **sexo** entre ambas fuentes de datos. 

El gráfico `escolar_plot` se utiliza para comparar el **nivel educativo** en los datos censales con los datos de la encuesta. 

Finalmente, el gráfico `depto_plot` compara las distribuciones de **entidad** entre los datos censales y de la encuesta. Estos gráficos permiten evaluar las diferencias en las características de la población entre las dos fuentes de datos.

```{r, eval=FALSE}
#### AGE ###
age_plot <-
  Plot_Compare(dat_censo = muestra_ampliada,
               dat_encuesta = encuesta_sta,
               by = "edad")
#### Sex ###
sex_plot <-
  Plot_Compare(dat_censo = muestra_ampliada,
               dat_encuesta = encuesta_sta,
               by = "sexo")
#### Level of schooling (LoS) ###
escolar_plot <-
  Plot_Compare(dat_censo = muestra_ampliada,
               dat_encuesta = encuesta_sta,
               by = "nivel_edu")

#### States ###
depto_plot <-
  Plot_Compare(dat_censo = muestra_ampliada,
               dat_encuesta = encuesta_sta,
               by = "ent")
```

Se crean gráficos comparativos para edad, sexo, nivel educativo y estados. Estos gráficos se combinan utilizando `patchwork`.

```{r, eval=FALSE}
#--- Patchwork en acción ---#
(age_plot | sex_plot | escolar_plot) / (depto_plot)
```

#### Comparación de Lengua Indígena y Discapacidad {-}

Para comparar las variables de **lengua indígena** y **discapacidad** entre los datos censales y los datos de la encuesta, se utilizan dos gráficos generados mediante la función `Plot_Compare`.

El gráfico `hlengua` muestra la comparación de la distribución de la variable **lengua indígena** en los datos censales frente a los datos de la encuesta. 

Por otro lado, el gráfico `plot_discapacidad` compara la distribución de la variable **discapacidad** entre ambos conjuntos de datos. 

Ambos gráficos se visualizan conjuntamente usando el operador `|`, permitiendo una comparación lado a lado de estas dos variables.

```{r, eval=FALSE}
hlengua <- Plot_Compare(dat_censo = muestra_ampliada,
               dat_encuesta = encuesta_sta,
               by = "hlengua")

plot_discapacidad <- Plot_Compare(dat_censo = muestra_ampliada,
             dat_encuesta = encuesta_sta,
             by = "discapacidad")

(plot_discapacidad | hlengua )
```

#### Efectos de Interacción {-}

Para analizar los efectos de interacción entre diferentes variables en los datos de la encuesta, se utiliza la función plot_interaction. A continuación, se presentan dos análisis específicos:

#### Interacción Edad x Sexo {-}

Se examina cómo varía el porcentaje de personas en situación de pobreza en función de la combinación de las variables **edad** y **sexo**. El gráfico `p_sex_age` ilustra esta interacción, mostrando cómo la pobreza se distribuye entre diferentes grupos etarios y de sexo.

```{r, eval=FALSE}
#### AGE x SEX ###
encuesta_sta$pobreza <- encuesta_sta$ictpc
#--- Percentage of people in poverty by AGE x SEX ---#
p_sex_age <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "edad")
```

#### Interacción Nivel Educativo x Sexo {-}
Se investiga la relación entre **nivel educativo** y **sexo** en cuanto a su influencia en la pobreza. El gráfico `p_sex_escolar` representa esta interacción, proporcionando una visión de cómo el nivel educativo combinado con el sexo afecta las tasas de pobreza.

```{r, eval=FALSE}
#### Level of schooling (LoS) x SEX ###
p_sex_escolar <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "nivel_edu")
```

#### Interacción Estado x Sexo {-}

Este análisis explora cómo varía la pobreza en función del **estado** y el **sexo**. El gráfico `p_sex_depto` muestra la distribución de la pobreza entre diferentes estados y géneros, proporcionando una visión sobre cómo estas variables combinadas influyen en la pobreza.

```{r, eval=FALSE}
#### State x SEX ###
p_sex_depto <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "ent")
```

Se combinan los gráficos de interacción de **Edad x Sexo** y **Nivel Educativo x Sexo** utilizando la funcionalidad `patchwork`. Esto permite una visualización conjunta para comparar fácilmente las interacciones entre las diferentes combinaciones de variables.

```{r, eval=FALSE}
#--- Patchwork in action ---#
(p_sex_age + p_sex_escolar) / p_sex_depto
```

#### Interacción Nivel Educativo x Edad {-}

Se analiza cómo el **nivel educativo** y la **edad** interactúan en relación con la pobreza. El gráfico `p_escolar_edad` muestra cómo varía el nivel educativo entre diferentes grupos etarios y se ajusta el tema para colocar la leyenda en la parte inferior.

```{r, eval=FALSE}
#### Level of schooling (LoS) x AGE ###
p_escolar_edad <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "nivel_edu",
                   by2 = "edad") +
  theme(legend.position = "bottom") + labs(colour = "nivel_edu")
```

#### Interacción Estado x Edad {-}

Se investiga la interacción entre el **estado** y la **edad** en cuanto a la pobreza. El gráfico `p_depto_edad` muestra la relación entre estos factores y se ajusta el tema para que la leyenda esté en la parte inferior.

```{r, eval=FALSE}
#### State x AGE ###
p_depto_edad <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "edad",
                   by2 = "ent") +
  theme(legend.position = "bottom") + labs(colour = "Edad")

p_escolar_edad / p_depto_edad
```

#### Interacción Nivel Educativo x Estado {-}
Se examina la relación entre el **nivel educativo** y el **estado** en términos de pobreza. El gráfico `p_depto_escolar` proporciona una visión de cómo el nivel educativo varía entre diferentes estados.

```{r, eval=FALSE}
#### Level of schooling (LoS) x State ###
p_depto_escolar <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "nivel_edu",
                   by2 = "ent") +
  theme(legend.position = "bottom") + labs(colour = "nivel_edu")

p_depto_escolar
```

#### Repetición del Análisis para ic_segsoc e ic_ali_nc {-}

Para realizar un análisis detallado de diferentes carencias, se repite el procedimiento de interacción para cada indicador específico. Este enfoque permite comparar cómo las interacciones entre variables afectan distintos aspectos de la pobreza. A continuación se presenta la descripción del análisis:


**Interacción Edad x Sexo:**
Se analiza la interacción entre **edad** y **sexo** en relación con el indicador de pobreza `ic_segsoc`. El gráfico `p_sex_age` muestra cómo varía la pobreza, medida por `ic_segsoc`, en función de la combinación de edad y sexo.

**Interacción Nivel Educativo x Sexo:**
Se examina cómo el **nivel educativo** interactúa con el **sexo** en relación con el indicador `ic_segsoc`. El gráfico `p_sex_escolar` proporciona una visión de la pobreza según el nivel educativo entre diferentes géneros.

**Interacción Estado x Sexo:**
El gráfico `p_sex_depto` muestra la pobreza medida por `ic_segsoc` en función del **estado** y el **sexo**, ayudando a entender cómo varía la pobreza entre diferentes estados y géneros.

**Combinación de Gráficos de Interacción:**
Se combinan los gráficos de interacción de **Edad x Sexo**, **Nivel Educativo x Sexo**, y **Estado x Sexo** usando `patchwork`, para una visualización consolidada de cómo estas combinaciones afectan el indicador `ic_segsoc`.

**Interacción Nivel Educativo x Edad:**
El gráfico `p_escolar_edad` muestra cómo el **nivel educativo** y la **edad** interactúan en relación con el indicador `ic_segsoc`. Se ajusta la leyenda para que esté en la parte inferior, facilitando la interpretación.

**Interacción Estado x Edad:**
Se explora la relación entre **edad** y **estado** en función del indicador `ic_segsoc`. El gráfico `p_depto_edad` ilustra cómo varía la pobreza entre diferentes edades y estados.

**Interacción Nivel Educativo x Estado:**
Finalmente, el gráfico `p_depto_escolar` muestra cómo el **nivel educativo** y el **estado** interactúan en relación con `ic_segsoc`, proporcionando una visión de la pobreza según el nivel educativo en distintos estados.

#### Resultados para ic_segsoc  {-}

```{r, eval=FALSE}
#### AGE x SEX ###
encuesta_sta$pobreza <- encuesta_sta$ic_segsoc
#--- Percentage of people in poverty by AGE x SEX ---#
p_sex_age <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "edad")

#### Level of schooling (LoS) x SEX ###
p_sex_escolar <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "nivel_edu")

#### State x SEX ###
p_sex_depto <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "ent")

#--- Patchwork in action ---#
(p_sex_age + p_sex_escolar) / p_sex_depto

#### Level of schooling (LoS) x AGE ###
p_escolar_edad <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "nivel_edu",
                   by2 = "edad") +
  theme(legend.position = "bottom") + labs(colour = "nivel_edu")

#### State x AGE ###
p_depto_edad <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "edad",
                   by2 = "ent") +
  theme(legend.position = "bottom") + labs(colour = "Edad")

p_escolar_edad / p_depto_edad

#### Level of schooling (LoS) x State ###
p_depto_escolar <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "nivel_edu",
                   by2 = "ent") +
  theme(legend.position = "bottom") + labs(colour = "nivel_edu")

p_depto_escolar
```

#### Resultados para ic_ali_nc {-}

```{r, eval=FALSE}
#### AGE x SEX ###
encuesta_sta$pobreza <- encuesta_sta$ic_ali_nc
#--- Percentage of people in poverty by AGE x SEX ---#
p_sex_age <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "edad")

#### Level of schooling (LoS) x SEX ###
p_sex_escolar <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "nivel_edu")

#### State x SEX ###
p_sex_depto <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "sexo",
                   by2 = "ent")

#--- Patchwork in action ---#
(p_sex_age + p_sex_escolar) / p_sex_depto

#### Level of schooling (LoS) x AGE ###
p_escolar_edad <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "nivel_edu",
                   by2 = "edad") +
  theme(legend.position = "bottom") + labs(colour = "nivel_edu")

#### State x AGE ###
p_depto_edad <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "edad",
                   by2 = "ent") +
  theme(legend.position = "bottom") + labs(colour = "Edad")

p_escolar_edad / p_depto_edad

#### Level of schooling (LoS) x State ###
p_depto_escolar <-
  plot_interaction(dat_encuesta = encuesta_sta,
                   by = "nivel_edu",
                   by2 = "ent") +
  theme(legend.position = "bottom") + labs(colour = "nivel_edu")

p_depto_escolar
```
